DATAES-629 - ElasticsearchEntityMapper refactoring.
Original PR: #309
