--- src/test/java/org/springframework/data/elasticsearch/core/convert/MappingElasticsearchConverterTests.java
@@ -1,64 +1,0 @@
-/*
- * Copyright 2013-2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.data.elasticsearch.core.convert;
-
-import static org.assertj.core.api.Assertions.*;
-
-import org.junit.Test;
-import org.springframework.core.convert.ConversionService;
-import org.springframework.data.elasticsearch.core.mapping.SimpleElasticsearchMappingContext;
-import org.springframework.data.mapping.context.MappingContext;
-
-/**
- * @author Rizwan Idrees
- * @author Mohsin Husen
- * @author Peter-Josef Meisch
- */
-public class MappingElasticsearchConverterTests {
-
-	@Test(expected = IllegalArgumentException.class)
-	public void shouldFailToInitializeGivenMappingContextIsNull() {
-
-		// given
-		new MappingElasticsearchConverter(null);
-	}
-
-	@Test
-	public void shouldReturnMappingContextWithWhichItWasInitialized() {
-
-		// given
-		MappingContext mappingContext = new SimpleElasticsearchMappingContext();
-		MappingElasticsearchConverter converter = new MappingElasticsearchConverter(mappingContext);
-
-		// then
-		assertThat(converter.getMappingContext()).isNotNull();
-		assertThat(converter.getMappingContext()).isSameAs(mappingContext);
-	}
-
-	@Test
-	public void shouldReturnDefaultConversionService() {
-
-		// given
-		MappingElasticsearchConverter converter = new MappingElasticsearchConverter(
-				new SimpleElasticsearchMappingContext());
-
-		// when
-		ConversionService conversionService = converter.getConversionService();
-
-		// then
-		assertThat(conversionService).isNotNull();
-	}
-}
