+++ src/test/java/org/springframework/data/elasticsearch/repository/support/ElasticsearchRepositoryFactoryTests.java
@@ -25,10 +25,7 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
 import org.springframework.data.elasticsearch.core.convert.ElasticsearchConverter;
 import org.springframework.data.elasticsearch.core.convert.MappingElasticsearchConverter;
-import org.springframework.data.elasticsearch.core.mapping.ElasticsearchPersistentEntity;
-import org.springframework.data.elasticsearch.core.mapping.ElasticsearchPersistentProperty;
 import org.springframework.data.elasticsearch.core.mapping.SimpleElasticsearchMappingContext;
-import org.springframework.data.mapping.context.MappingContext;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.data.repository.core.RepositoryMetadata;
 import org.springframework.data.repository.core.support.DefaultRepositoryMetadata;
@@ -37,19 +34,17 @@ import org.springframework.data.repository.core.support.DefaultRepositoryMetadat
  * @author Rizwan Idrees
  * @author Mohsin Husen
  * @author Mark Paluch
+ * @author Peter-Josef Meisch
  */
 @RunWith(MockitoJUnitRunner.class)
 public class ElasticsearchRepositoryFactoryTests {
 
-	@Mock private ElasticsearchOperations operations;
-	private ElasticsearchConverter converter;
+    @Mock private ElasticsearchOperations operations;
 	private ElasticsearchRepositoryFactory factory;
-	MappingContext<? extends ElasticsearchPersistentEntity<?>, ElasticsearchPersistentProperty> mappingContext = new SimpleElasticsearchMappingContext();
 
 	@Before
 	public void before() {
-
-		converter = new MappingElasticsearchConverter(mappingContext);
+        ElasticsearchConverter converter = new MappingElasticsearchConverter(new SimpleElasticsearchMappingContext());
 		when(operations.getElasticsearchConverter()).thenReturn(converter);
 		factory = new ElasticsearchRepositoryFactory(operations);
 	}
