--- src/test/resources/elasticsearch-template-custom-mapper.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans https://www.springframework.org/schema/beans/spring-beans-3.1.xsd">
-
-    <import resource="infrastructure.xml"/>
-
-    <bean name="elasticsearchTemplate"
-          class="org.springframework.data.elasticsearch.core.ElasticsearchTemplate">
-        <constructor-arg name="client" ref="client"/>
-        <constructor-arg name="resultsMapper" ref="resultMapper"/>
-    </bean>
-
-    <bean name="entityMapper" class="org.springframework.data.elasticsearch.core.CustomEntityMapper"/>
-
-    <bean name="resultMapper" class="org.springframework.data.elasticsearch.core.CustomResultMapper">
-        <constructor-arg ref="entityMapper"/>
-    </bean>
-
-</beans>
